Decrypt for same path bug fix

+++ src/main/java/com/mastercard/developer/encryption/FieldLevelEncryption.java
@@ -193,10 +193,12 @@ public class FieldLevelEncryption {
             JsonParser.checkOrCreateOutObject(payloadContext, jsonPathOut);
             JsonParser.addDecryptedDataToPayload(payloadContext, decryptedValue, jsonPathOut);
 
-            // Remove the input if now empty
-            Object inJsonElement  = JsonParser.readJsonElement(payloadContext, jsonPathIn);
-            if (0 == jsonProvider.length(inJsonElement)) {
-                payloadContext.delete(jsonPathIn);
+            if(jsonPathIn != jsonPathOut) {
+                // Remove the input if now empty
+                Object inJsonElement  = JsonParser.readJsonElement(payloadContext, jsonPathIn);
+                if (0 == jsonProvider.length(inJsonElement)) {
+                    payloadContext.delete(jsonPathIn);
+                }
             }
         }
